requirements.txt
aiogram>=3.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
